+++ udata/core/dataset/models.py
@@ -115,8 +115,10 @@ class License(db.Document):
     id = db.StringField(primary_key=True)
     created_at = db.DateTimeField(default=datetime.now, required=True)
     title = db.StringField(required=True)
+    alternate_titles = db.ListField(db.StringField())
     slug = db.SlugField(required=True, populate_from='title')
     url = db.URLField()
+    alternate_urls = db.ListField(db.URLField())
     maintainer = db.StringField()
     flags = db.ListField(db.StringField())
 
@@ -153,10 +155,25 @@ class License(db.Document):
         qs = cls.objects
         text = text.strip().lower()  # Stored identifiers are lower case
         slug = cls.slug.slugify(text)  # Use slug as it normalize string
-        license = qs(db.Q(id=text) | db.Q(slug=slug) | db.Q(url=text)).first()
+        license = qs(
+            db.Q(id=text) | db.Q(slug=slug) | db.Q(url=text)
+            | db.Q(alternate_urls=text)
+        ).first()
         if license is None:
             # Try to single match with a low Damerau-Levenshtein distance
             computed = ((l, rdlevenshtein(l.slug, slug)) for l in cls.objects)
+            candidates = [l for l, d in computed if d <= MAX_DISTANCE]
+            # If there is more that one match, we cannot determinate
+            # which one is closer to safely choose between candidates
+            if len(candidates) == 1:
+                license = candidates[0]
+        if license is None:
+            # Try to single match with a low Damerau-Levenshtein distance
+            computed = (
+                (l, rdlevenshtein(cls.slug.slugify(t), slug))
+                for l in cls.objects
+                for t in l.alternate_titles
+            )
             candidates = [l for l, d in computed if d <= MAX_DISTANCE]
             # If there is more that one match, we cannot determinate
             # which one is closer to safely choose between candidates