+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [breaking] Enforce a domain whitelist when resource.filetype is file [#1567](https://github.com/opendatateam/udata/issues/1567)
 - Initial data preview implementation [#1581](https://github.com/opendatateam/udata/pull/1581)
 - Switch to PyPI.org for package links [#1583](https://github.com/opendatateam/udata/pull/1583)
+- Handle some alternate titles and alternate URLs on licenses for improved match on harvesting [#1592](https://github.com/opendatateam/udata/pull/1592)
 
 ## 1.3.6 (2018-04-16)